Move generate simulation configs to neuron subdirectory
