+++ obi_one/scientific/tasks/generate_simulations/config/neuron/me_model_with_synapses.py
@@ -13,7 +13,9 @@ from obi_one.scientific.tasks.generate_simulations.config.base import (
     SimulationScanConfig,
     SimulationSingleConfigMixin,
 )
-from obi_one.scientific.tasks.generate_simulations.config.circuit import CircuitSimulationScanConfig
+from obi_one.scientific.tasks.generate_simulations.config.neuron.circuit import (
+    CircuitSimulationScanConfig,
+)
 from obi_one.scientific.unions.unions_neuron_sets import (
     MEModelWithSynapsesNeuronSetUnion,
     NeuronSetReference,
