+++ tests/obi_one/scientific/tasks/ion_channel_model_simulation_execution/test_ion_channel_model_simulation_execution.py
@@ -8,7 +8,7 @@ from obi_one.scientific.blocks.ion_channel_model import IonChannelModelWithCondu
 from obi_one.scientific.from_id.ion_channel_model_from_id import IonChannelModelFromID
 from obi_one.scientific.library.simulation.schemas import SimulationParameters, SimulationResults
 from obi_one.scientific.tasks import ion_channel_model_simulation_execution as test_module
-from obi_one.scientific.tasks.generate_simulations.config.ion_channel_models import (
+from obi_one.scientific.tasks.generate_simulations.config.neuron.ion_channel_models import (
     IonChannelModelSimulationScanConfig,
     IonChannelModelSimulationSingleConfig,
 )
